chore: typescript refactor
