+++ js/src/forum/extenders/addFrontSort.ts
@@ -1,8 +1,10 @@
 import { extend } from 'flarum/common/extend';
 import DiscussionListState from 'flarum/forum/states/DiscussionListState';
 
+import type { PaginatedListParams } from 'flarum/common/states/PaginatedListState';
+
 export default function () {
-  extend(DiscussionListState.prototype, 'requestParams', function (params) {
+  extend(DiscussionListState.prototype, 'requestParams', function (this: DiscussionListState, params: PaginatedListParams) {
     if (this.params.sort === 'front') {
       if (params.filter.q) {
         params.filter.q = (params.filter.q || '') + 'is:frontpage';
@@ -12,7 +14,7 @@ export default function () {
     }
   });
 
-  extend(DiscussionListState.prototype, 'sortMap', function (map) {
+  extend(DiscussionListState.prototype, 'sortMap', function (map: Record<string, string>) {
     map.front = '-frontdate';
   });
 }
